arch: arm: dts: ipq5332: configure SFP RX pin state
We used to configure SFP RX pin with 8MA drive strenght and pull
up enabled, but later identified that, it is not really required.
Then, removed it in miami.

But, looks likes when device like SFP-80M module is connected to
SFP cage, these configurations are needed. So, enabling it on
all miami sfp rdps.

This change wont affect our SFP-30M module behaviour.

Change-Id: I5dd758e1d982062d9a0259b50e1e537eeda8ddc8
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>